Move XML serialization to TestUtils
