+++ BabelWiresLib/CMakeLists.txt
@@ -153,7 +153,7 @@ TARGET_INCLUDE_DIRECTORIES( BabelWiresLib
 		$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>
 		$<INSTALL_INTERFACE:include>
 )
-TARGET_LINK_LIBRARIES(BabelWiresLib PUBLIC tinyxml2 BaseLib)
+TARGET_LINK_LIBRARIES(BabelWiresLib PUBLIC BaseLib)
 
 if(BUILD_SHARED_LIBS)
 	set_target_properties(BabelWiresLib PROPERTIES
