+++ BaseLib/CMakeLists.txt
@@ -27,8 +27,6 @@ SET( COMMON_SRCS
 	Serialization/serializerDeserializerCommon.cpp
 	Serialization/deserializationRegistry.cpp
 	Serialization/userDocumentSerializationFactory.cpp
-	Serialization/XML/xmlSerializer.cpp
-	Serialization/XML/xmlDeserializer.cpp
 	Serialization/YAML/yamlSerializer.cpp
 	Serialization/YAML/yamlDeserializer.cpp
 	Signal/signalSubscription.cpp
@@ -60,7 +58,7 @@ TARGET_INCLUDE_DIRECTORIES( BaseLib
 		$<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/generated>
 		$<INSTALL_INTERFACE:include>
 )
-TARGET_LINK_LIBRARIES(BaseLib PUBLIC tinyxml2 yaml-cpp)
+TARGET_LINK_LIBRARIES(BaseLib PUBLIC yaml-cpp)
 
 if(UNIX AND NOT APPLE)
 	target_link_libraries(BaseLib PRIVATE dl)
