+++ Tests/TestUtils/CMakeLists.txt
@@ -3,8 +3,10 @@ SET( TEST_UTILS_SRCS
    tempFilePath.cpp
    testIdentifiers.cpp
    testStrings.cpp
+   XmlSerialization/xmlSerializer.cpp
+   XmlSerialization/xmlDeserializer.cpp
    )
 
 ADD_LIBRARY( testUtils STATIC ${TEST_UTILS_SRCS} )
 TARGET_INCLUDE_DIRECTORIES( testUtils PRIVATE ${PROJECT_SOURCE_DIR} )
-TARGET_LINK_LIBRARIES(testUtils BaseLib)
+TARGET_LINK_LIBRARIES(testUtils PUBLIC BaseLib tinyxml2)
