+++ Tests/TestUtils/XmlSerialization/xmlDeserializer.hpp
@@ -2,12 +2,10 @@
  * The XmlDeserializer implements the Deserializer and loads data from an XML representation.
  *
  * (C) 2021 Malcolm Tyrrell
- * 
+ *
  * Licensed under the GPLv3.0. See LICENSE file.
  **/
 #pragma once
-
-#include <BaseLib/baseLibExport.hpp>
 
 #include <BaseLib/Serialization/deserializer.hpp>
 
@@ -16,7 +14,7 @@
 
 namespace babelwires {
 
-    class BASELIB_API XmlDeserializer : public Deserializer {
+    class XmlDeserializer : public Deserializer {
       public:
         XmlDeserializer(const DeserializationRegistryInterface& deserializationRegistry, UserLogger& userLogger);
 
@@ -35,7 +33,7 @@ namespace babelwires {
         ResultT<bool> tryDeserializeValue(std::string_view key, std::int8_t& value) override;
 
       protected:
-        struct BASELIB_API IteratorImpl : Deserializer::AbstractIterator {
+        struct IteratorImpl : Deserializer::AbstractIterator {
             IteratorImpl(XmlDeserializer& deserializer, const tinyxml2::XMLElement* arrayElement);
 
             Result advance() override;
